feat(order): 新增订单创建和查询功能
- 新增订单创建逻辑，支持商品快照和分类快照
- 实现订单列表查询功能，返回订单详情和商品信息
- 优化购物车和结算页面，更新商品展示和价格计算
- 新增 DevPlan.md 文件，记录项目开发计划和进度

+++ src/main/java/com/jiang/mall/controller/OrderController.java
@@ -198,7 +198,7 @@ public class OrderController {
 	        return result;
 	    }
 	    // 检查 session.getAttribute("List_cartId") 是否为 null
-	    List<Integer> list_cartId;
+	    List<Long> list_cartId;
 	    Object listObj = session.getAttribute("List_cartId");
 
 	    if (listObj == null) {
@@ -206,8 +206,8 @@ public class OrderController {
 	    } else if (listObj instanceof List<?> tempList) {
 	        list_cartId = new ArrayList<>();
 	        for (Object obj : tempList) {
-	            if (obj instanceof Integer) {
-	                list_cartId.add((Integer) obj);
+	            if (obj instanceof Long) {
+	                list_cartId.add((Long) obj);
 	            }
 	        }
 	    } else {
